Tarea 2/elastic/requirements.txt
kafka-python
elasticsearch

